fixed? problem with alarms going off twice. improved UI stuff. integrated alarmclock further. rating now shows up when reviewing sleepcharviews. cleaned up and added force-calibration to the welcome wizard.
this is the last commit before beta refactoring...

+++ src/com/androsz/electricsleep/app/SaveSleepActivity.java
@@ -49,7 +49,7 @@ public class SaveSleepActivity extends CustomTitlebarActivity implements
 			return;
 		}
 
-		getIntent().putExtra("rating", rating);
+		getIntent().putExtra("rating", (int)rating);
 
 		final Intent saveIntent = new Intent(SaveSleepActivity.SAVE_SLEEP);
 		saveIntent.putExtras(getIntent().getExtras());
